# Improve (hugely!) impact graph performance