--- a/Plugins/Statistics/GitImpact/ImpactLoader.cs
+++ b/Plugins/Statistics/GitImpact/ImpactLoader.cs
@@ -75,6 +75,7 @@ namespace GitExtensions.Plugins.GitImpact
             _closingPluginCancellationToken.Cancel();
             Stop();
             _cancellationTokenSequence.Dispose();
+            _closingPluginCancellationToken.Dispose();
         }
 
         public void Stop()