<!-- Please read CONTRIBUTING.md before submitting a pull request -->

Fixes #11411

## Proposed changes

1. Rendering now takes less than 1s (whereas it was taking minutes on big repositories like GitExtensions --around 7min-- or Git --around 1h08min--ones).

So now the time to build the impact graph correspond only to the time to retrieve data from repo. 

I did one benchmark on git repository for the rendering:
* before: 1h8min
* after: 0.76s (around x5300 times quicker!!)

Retrieving data still takes around 30s, so overall process was before 1h9min and after 30s (x130 times quicker!)

GitExtensions repo: now total time 23s (200ms of rendering)

2. Submodule checkbox change doesn't cancel the long main data retrieve and data are put in a cache so that switch is immediate (instead of fetching the data again every times)
3. Display is done depending on 1st day of the week culture setting


## Test methodology <!-- How did you ensure quality? -->

- Manual

## Test environment(s) <!-- Remove any that don't apply -->

- Git Extensions 33.33.33
- Build 72f0d419fcce299752cc26c96009c6dab53aba89 (Dirty)
- Git 2.42.0.windows.2
- Microsoft Windows NT 10.0.22621.0
- .NET 8.0.0
- DPI 96dpi (no scaling)
- Portable: False
- Microsoft.WindowsDesktop.App Versions

## Merge strategy

<!-- Change the following if the merge strategy should be changed:
- Squash merge (maintainer to decide merge message, PR submitter should cleanup commits/messages at PR approval).
- Rebase merge (PR submitter must change the commit message for the last commit).
- Merge commit. (PR submitter to rebase and squash before merges).
- To be decided later.
The maintainer may still request the contributor to rebase (preferred) or squash and rebase, to make sure that merges and commit messages are clarified.
-->

I agree that the maintainer rebase (preferred) or squash merge this PR (if the commit message is clear).

----

:black_nib: I contribute this code under [The Developer Certificate of Origin](../blob/master/contributors.txt).