@pmiossec 2023-12-20
I think I fixed all the review issue and done an overall pass on disposing resources.
Since the last review, I also took the opportunity to do 2 little things: converting usage to `DateOnly` and adapting display depending on "1st day of the week" culture preference.